requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
nltk
python-dotenv
groq
pdfplumber
pytesseract
Pillow